create-gaarf-wf@1.6.0: - feat: added a check for that a path to google-ads.yaml really leads to a file - feat: added auto-sanitizing of dataset names (replacing dashes with underscores) - feat: login_customer_id from google-ads.yaml is used as default CID value - feat: added a check to wrap job deletion in schedule-job.sh - feat: added an ability to run workflow synchronously

+++ gcp/create-gaarf-wf/build/index.js
@@ -311,7 +311,10 @@ function get_macro_values(folder_path, answers, prefix) {
             options.filter(i => !(i.name in answers[prefix])).length) {
             console.log(`Please enter values for the following macros found in your scripts in '${folder_path}' folder`);
             console.log(chalk.yellow('Tip: ') +
-                chalk.gray('beside constants you can use :YYYYMMDD-N values and expressions (${..})'));
+                chalk.gray('besides constants you can use :YYYYMMDD-N values and expressions (${..})') +
+                '\n' +
+                chalk.yellow('Tip: ') +
+                chalk.gray('For macros with dataset names make sure that you meet BigQuery requirements - use letters, numbers and underscores'));
         }
         answers[prefix] = answers[prefix] || {};
         return prompt(options, answers[prefix]);
@@ -448,12 +451,15 @@ async function initialize_googleads_config(answers) {
             },
         ], answers);
         path_to_googleads_config = answers_new.path_to_googleads_config;
-        const has_adsconfig = fs.existsSync(path_to_googleads_config);
-        if (!has_adsconfig) {
+        if (!fs.existsSync(path_to_googleads_config)) {
             console.log(chalk.yellow('Currently the file ') +
                 chalk.cyan(path_to_googleads_config) +
                 chalk.yellow("does not exist, please note that you need to upload it before you can actually deploy and run, after that you'll need to run ") +
                 chalk.cyan('deploy-scripts.sh'));
+        }
+        else if (!fs.statSync(path_to_googleads_config).isFile()) {
+            console.log(chalk.red('The path to google-ads.yaml you specified does not exist. You can a full or relative file path but it should include a file name'));
+            process.exit(-1);
         }
     }
     else {
@@ -718,14 +724,21 @@ cd ../workflow
     let ads_customer_id;
     if (fs.existsSync(path_to_googleads_config)) {
         const yamldoc = (yaml.load(fs.readFileSync(path_to_googleads_config, 'utf-8')));
-        ads_customer_id = yamldoc['customer_id'] || yamldoc['client_customer_id'];
+        // look up for the default default value for account id (CID) from google-ads.yaml
+        ads_customer_id =
+            yamldoc['customer_id'] ||
+                yamldoc['client_customer_id'] ||
+                yamldoc['login_customer_id'];
     }
     const answers2 = await prompt([
         {
             type: 'input',
             name: 'output_dataset',
-            message: 'BigQuery dataset for ads queries results:',
+            message: 'BigQuery dataset for ads queries results ("-" will be converted to "_"):',
             default: name + '_ads',
+            filter(input) {
+                return input.replace(/[ -]/g, '_');
+            },
         },
         {
             type: 'input',
@@ -757,7 +770,7 @@ cd ../workflow
     };
     // Create run-wf.sh
     deploy_shell_script('run-wf.sh', `gcloud workflows run ${workflow_name} --location=${gcp_region} \
-  --data='${JSON.stringify(wf_data, null, 2)}'
+--data='${JSON.stringify(wf_data, null, 2)}'
 `);
     // now execute some scripts
     if (ready_to_deploy_scripts) {
@@ -810,7 +823,7 @@ cd ../workflow
             {
                 type: 'confirm',
                 name: 'run_job',
-                message: 'Do you want to run the job right now:',
+                message: "Do you want to run the job right now (it's asynchronous):",
             },
         ], answers);
         const time_parts = answers_schedule.schedule_time.split(':');
@@ -830,9 +843,12 @@ JOB_NAME=$WORKFLOW_NAME
 
 data='${JSON.stringify(wf_data, null, 2).replaceAll('"', '\\"')}'
 
-gcloud scheduler jobs delete $JOB_NAME --location $REGION --quiet
+JOB_EXISTS=$(gcloud scheduler jobs list --location=$REGION --format="value(ID)" --filter="ID:'$JOB_NAME'")
+if [[ -n $JOB_EXISTS ]]; then
+  gcloud scheduler jobs delete $JOB_NAME --location $REGION --quiet
+fi
 
-# daily at midnight
+# run the job daily at midnight
 gcloud scheduler jobs create http $JOB_NAME \\
   --schedule="${schedule_cron}" \\
   --uri="https://workflowexecutions.googleapis.com/v1/projects/$PROJECT_ID/locations/$REGION/workflows/$WORKFLOW_NAME/executions" \
@@ -856,6 +872,19 @@ gcloud scheduler jobs create http $JOB_NAME \\
         if (answers3.run_job) {
             // running the job
             await exec_cmd(`gcloud scheduler jobs run ${workflow_name} --location=${gcp_region}`, null, { realtime: true });
+        }
+    }
+    if (!answers3.run_job) {
+        // Scheduler Job wasn't run, maybe the user want to run the workflow directly (it's synchronous in contract to the scheduler)
+        const answers_wf = await prompt([
+            {
+                type: 'confirm',
+                name: 'run_wf',
+                message: "Do you want to run the workflow right now (it's synchronous):",
+            },
+        ], answers);
+        if (answers_wf.run_wf) {
+            await exec_cmd(path.join(cwd, './run-wf.sh'), null, { realtime: true });
         }
     }
     // creating scripts for directly executing gaarf
